Product-fix

+++ Banco/src/main/java/com/reto/Banco/controller/clienteController.java
@@ -3,14 +3,17 @@ package com.reto.Banco.controller;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import org.hibernate.annotations.Any;
 import org.springframework.beans.factory.annotation.Autowired;
-
+import org.springframework.boot.autoconfigure.amqp.RabbitProperties.Retry;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.Trigger;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,11 +24,15 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.reto.Banco.dto.GeneralResponse;
 import com.reto.Banco.entity.ClientTable;
+import com.reto.Banco.entity.ProductEntity;
 import com.reto.Banco.service.ClientService;
+import com.reto.Banco.service.ProductSevice;
+
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
 import java.io.Console;
+import java.text.BreakIterator;
 import java.text.SimpleDateFormat;  
 
 @CrossOrigin(origins = "*")
@@ -35,7 +42,8 @@ public class clienteController {
 
     @Autowired
     ClientService clientService;
-
+    @Autowired 
+    ProductSevice productService;
     /**
      * @return
      */
@@ -130,7 +138,7 @@ public class clienteController {
             // Date date1 = new SimpleDateFormat("dd/MM/yyyy").parse(client.getAuxBirthdate());          
             // client.setBirthdate(date1);  
             // System.out.println("Test: "+client.getAuxBirthdate());
-            System.out.println(client.getIdNum());
+            // System.out.println(client.getIdNum());
 
             finalClient = new ClientTable(            client.getId_Type(),
                                                       client.getIdNum(),
@@ -150,21 +158,23 @@ public class clienteController {
                                                       
 
         //  finalClient.setAuxBirthdate(client.getAuxBirthdate());
-            
+            long years = ChronoUnit.YEARS.between( client.getBirthdate(),  LocalDate.now());
+           
+            System.out.println(years > 17);
             //restar edades
-            if(/*client.getAge() > 18*/ true )
+            if (years > 17 )
             {
                 // client.setDatecreation(LocalDate.now());
                 datos = clientService.CreateCliente(finalClient);
+                respuesta.setDatos(finalClient);
                 mensaje = "0 - Customer successfully created";
             }else {
                 mensaje ="1 - Customer could not be create, the age ";
-				estadoHttp = HttpStatus.OK;
+				estadoHttp = HttpStatus.BAD_REQUEST;
             }
             
 
 
-            respuesta.setDatos(finalClient);
             respuesta.setMensaje(mensaje);
             respuesta.setPeticionExitosa(true);
             estadoHttp = HttpStatus.CREATED;
@@ -234,17 +244,48 @@ public class clienteController {
         GeneralResponse<Long> respuesta = new GeneralResponse<>();        
 		String mensaje = null;	
 		HttpStatus estadoHttp = null;    
-
+        List<ProductEntity> Products; 
+        boolean trigger = false;
+        
         try{
+            
+            Products =  productService.findByclienteId(id);
+            // System.out.println(Products.size());
+
+
+            
+            
+            for(int i = 0 ; i < Products.size()-1; i++ )
+            {
+                if(Products.get(i).getState().equalsIgnoreCase("cancelled") )
+                {
+                    // System.out.println(!Products.get(i).getEstado().equalsIgnoreCase("cancelled"));
+                    trigger = true;
+                    break;
+                }
+            }
+            System.out.println(trigger);
+
+            if( trigger == false)
+            {
+                clientService.delete(id);
+                mensaje = "0 - Customer successfully delete";
+                respuesta.setPeticionExitosa(true);
+                estadoHttp = HttpStatus.ACCEPTED;
+            }else 
+            {
+                mensaje = "1 - Customer don't be  delete, exist "+ Products.size() + " active product.";
+                respuesta.setPeticionExitosa(false);
+                estadoHttp = HttpStatus.BAD_REQUEST;
+            }
+
+
+
             // client.setId(id);
-            clientService.delete(id);
            
-            mensaje = "0 - Customer successfully delete";
 
             respuesta.setDatos(id);
             respuesta.setMensaje(mensaje);
-            respuesta.setPeticionExitosa(true);
-            estadoHttp = HttpStatus.CREATED;
         }catch(Exception e){
 
             estadoHttp = HttpStatus.INTERNAL_SERVER_ERROR;
